+++ firmware/Quaverato/src/relay.cpp
@@ -3,7 +3,19 @@
 #include <Arduino.h>
 #include "pins.h"
 #include "state.h"
-#include "tasks.h"
+
+static unsigned long relayTwoAt = 0;
+static unsigned long relayCleanupAt = 0;
+static unsigned long relayIsolatorAt = 0;
+static unsigned long midiMomentAt = 0;
+static bool relayTwoArmed = false;
+static bool relayCleanupArmed = false;
+static bool relayIsolatorArmed = false;
+static bool midiMomentArmed = false;
+
+static bool deadlineReached(unsigned long at) {
+  return (long)(micros() - at) >= 0;
+}
 
 void flipRelay() {
   static unsigned long localButtonTimer = 0;
@@ -11,7 +23,8 @@ void flipRelay() {
 
   if (micros() - localButtonTimer > 25000) {
     digitalWrite(relay_pin_Isolator, LOW);
-    relayPartTwo.enableDelayed(25000);
+    relayTwoAt = micros() + 25000;
+    relayTwoArmed = true;
   }
   localButtonTimer = micros();
   interrupts();
@@ -24,25 +37,47 @@ void flipRelayPartTwo() {
   digitalWrite(relay_pin_Switch, LOW);
   digitalWrite(relay_pin_Lock, !relayON);
   digitalWrite(relay_pin_Switch, relayON);
-  relayCleanup.enableDelayed(8000);
-  relayPartTwo.disable();
+  relayCleanupAt = micros() + 8000;
+  relayCleanupArmed = true;
+  relayTwoArmed = false;
 }
 
 void resetRelay() {
   digitalWrite(relay_pin_Lock, LOW);
   digitalWrite(relay_pin_Switch, LOW);
-  relayDisableIsolator.enableDelayed(5000);
-  relayCleanup.disable();
+  relayIsolatorAt = micros() + 5000;
+  relayIsolatorArmed = true;
+  relayCleanupArmed = false;
 }
 
 void disableIsolator() {
   digitalWrite(relay_pin_Isolator, HIGH);
-  relayDisableIsolator.disable();
+  relayIsolatorArmed = false;
+}
+
+void armMidiMomentary() {
+  midiMomentAt = micros() + 20000;
+  midiMomentArmed = true;
 }
 
 void midiMomentRelay() {
   if (!momentMode) {
     flipRelay();
   }
-  midiMomentaryMode.disable();
+  midiMomentArmed = false;
+}
+
+void serviceRelay() {
+  if (relayTwoArmed && deadlineReached(relayTwoAt)) {
+    flipRelayPartTwo();
+  }
+  if (relayCleanupArmed && deadlineReached(relayCleanupAt)) {
+    resetRelay();
+  }
+  if (relayIsolatorArmed && deadlineReached(relayIsolatorAt)) {
+    disableIsolator();
+  }
+  if (midiMomentArmed && deadlineReached(midiMomentAt)) {
+    midiMomentRelay();
+  }
 }