+++ firmware/Quaverato/src/state.cpp
@@ -5,8 +5,8 @@ const int minTapRate = 784;   // actual minimum is 784/4 = 196
 const int maxTapRate = 15625; // actual maximum is 15625*2 = 31250 microseconds
 
 unsigned long stepRate = 196;
-unsigned long firstHalfStepRate = 196;
-unsigned long secondHalfStepRate = 196;
+volatile unsigned long firstHalfStepRate = 196;
+volatile unsigned long secondHalfStepRate = 196;
 float dutyCycle = 0.5;
 
 volatile bool relayON = false;
@@ -15,8 +15,8 @@ volatile unsigned long debounceTimer = 0;
 volatile unsigned long buttonTimer = 0;
 volatile unsigned long tempoTapped = 0;
 
-byte currentWaveTable = 0;
-byte waveFormStep = 0;
+volatile byte currentWaveTable = 0;
+volatile byte waveFormStep = 0;
 
 byte depth = 16;
 double tapDivisor = 1;
@@ -37,7 +37,7 @@ bool tableShift = false;
 bool setMidiChannel = false;
 bool midiOmni = false;
 int midiChannel = 0;
-bool followMidiClock = false;
+volatile bool followMidiClock = false;
 int clockCount = 0;
 
 bool momentMode = false;