+++ firmware/Quaverato/src/ts.cpp
@@ -1,5 +1,0 @@
-// Sole translation unit that includes the full TaskScheduler implementation.
-// Do not include other project headers here.
-
-#define _TASK_MICRO_RES
-#include <TaskScheduler.h>